Split resolver behavior metadata tests
