+++ src/typechecker/tests/resolver_type_behavior_metadata/behavior_methods.rs
@@ -1,0 +1,181 @@
+use super::*;
+
+#[test]
+fn check_program_with_symbols_validates_resolver_behavior_method_signatures() {
+    let program = parse_program(
+        r#"
+Serializable: behavior {
+    encode: (Self, i32) str
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_method_signatures_for_test(
+        Namespace::Behavior,
+        "Serializable",
+        Some(vec![(
+            "encode".to_string(),
+            vec!["Self".to_string(), "bool".to_string()],
+            "str".to_string(),
+        )]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver behavior method signature mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver behavior symbol 'Serializable' has methods '(encode(Self, bool) str)', expected '(encode(Self, i32) str)'"
+            )),
+            "expected resolver behavior method signature diagnostic, got {err:?}"
+        );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_behavior_function_type_method_signatures() {
+    let program = parse_program(
+        r#"
+Mapper: behavior {
+    map: (Self, (i32) i32) (i32) i32
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_method_signatures_for_test(
+        Namespace::Behavior,
+        "Mapper",
+        Some(vec![(
+            "map".to_string(),
+            vec!["Self".to_string(), "i32".to_string()],
+            "(i32) i32".to_string(),
+        )]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver behavior function type method signature mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver behavior symbol 'Mapper' has methods '(map(Self, i32) (i32) i32)', expected '(map(Self, (i32) i32) (i32) i32)'"
+            )),
+            "expected resolver behavior function type method signature diagnostic, got {err:?}"
+        );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_behavior_method_types() {
+    let program = parse_program(
+        r#"
+Mapper: behavior {
+    map: (Self, (i32) i32) (i32) i32
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_method_types_for_test(
+        Namespace::Behavior,
+        "Mapper",
+        Some(vec![BehaviorMethodTypeMetadata {
+            name: "map".to_string(),
+            parameter_names: vec!["__arg0".to_string(), "__arg1".to_string()],
+            parameter_types: vec![AstType::SelfType, AstType::I32],
+            return_type: AstType::Function {
+                params: vec![AstType::I32],
+                ret: Box::new(AstType::I32),
+            },
+        }]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver typed behavior method metadata mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver behavior symbol 'Mapper' has typed methods '(map(__arg0: Self, __arg1: i32) (i32) i32)', expected '(map(__arg0: Self, __arg1: (i32) i32) (i32) i32)'"
+            )),
+            "expected resolver typed behavior method diagnostic, got {err:?}"
+        );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_generic_behavior_method_signatures() {
+    let program = parse_program(
+        r#"
+Json<T>: behavior {
+    encode: (Self) T
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_method_signatures_for_test(
+        Namespace::Behavior,
+        "Json",
+        Some(vec![(
+            "encode".to_string(),
+            vec!["Self".to_string()],
+            "str".to_string(),
+        )]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver generic behavior method signature mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver behavior symbol 'Json' has methods '(encode(Self) str)', expected '(encode(Self) T)'"
+            )),
+            "expected resolver generic behavior method signature diagnostic, got {err:?}"
+        );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_generic_behavior_function_type_method_signatures()
+{
+    let program = parse_program(
+        r#"
+Mapper<T>: behavior {
+    map: (Self, (T) T) (T) T
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_behavior_method_signatures_for_test(
+        Namespace::Behavior,
+        "Mapper",
+        Some(vec![(
+            "map".to_string(),
+            vec!["Self".to_string(), "T".to_string()],
+            "(T) T".to_string(),
+        )]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver generic behavior function type method mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver behavior symbol 'Mapper' has methods '(map(Self, T) (T) T)', expected '(map(Self, (T) T) (T) T)'"
+            )),
+            "expected resolver generic behavior function type method diagnostic, got {err:?}"
+        );
+}
